# packages/contract/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arch-Systems Portal API",
    "version": "1.0.0",
    "description": "API for industrial operations portal - control room, drilling, engineering, safety, and production management"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT authentication token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/sync/playback": {
      "post": {
        "summary": "Queue sync playback event",
        "description": "Queue a sync playback event via Inngest for offline-first data synchronization. Uses idempotency keys to prevent duplicate processing.",
        "tags": [
          "Sync"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey",
                  "actionType",
                  "payload",
                  "departmentId"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "description": "Unique key to prevent duplicate processing"
                  },
                  "actionType": {
                    "type": "string",
                    "description": "Type of sync action (create, update, delete)"
                  },
                  "payload": {
                    "type": "object",
                    "description": "Action payload data"
                  },
                  "departmentId": {
                    "type": "string",
                    "description": "Department ID for the sync event"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event queued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "queued": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing required fields"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/control-room/shift-completeness": {
      "get": {
        "summary": "Get shift completeness metrics",
        "description": "Returns completeness metrics for a specific department, date, and shift. Requires authentication.",
        "tags": [
          "Control Room"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "deptId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Department UUID"
          },
          {
            "in": "query",
            "name": "deptSlug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Department slug"
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Shift date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "shift",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "night"
              ]
            },
            "description": "Shift type"
          }
        ],
        "responses": {
          "200": {
            "description": "Shift completeness data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}